+++ terminal/src/main/java/it/units/sdm/project/TerminalLauncher.java
@@ -1,6 +1,5 @@
 package it.units.sdm.project;
 
-import it.units.sdm.project.core.game.terminal.FreedomGame;
 import it.units.sdm.project.core.board.MapBoard;
 import it.units.sdm.project.core.board.Stone;
 import it.units.sdm.project.core.game.Player;
